feat(blog): detail
- Ajout de la pagination dans l'admin
- Ajout de la pagination dans le front
- Ajout de la propriété "/ext" dans l'entité Document.php
- Modification de tout les controllers pour gérer le nouveau system d'extension
- Modification du FileService.php pour gérer les extensions
- Adaptation de toute les vues pour gérer les extensions

+++ src/Controller/BlogController.php
@@ -17,15 +17,17 @@ use Symfony\Component\Routing\Annotation\Route;
 class BlogController extends AbstractController
 {
     /**
-     * @Route("/blog", name="blog_show_list", methods={"GET"})
+     * @Route("/blog", defaults={"page": "1"}, name="blog_show_list", methods={"GET"})
+     * @Route("/blog/{page<[1-9]\d*>}", methods="GET", name="blog_show_list_paginated")
      * @param BlogRepository $blogRepository
      * @param CategoryRepository $categoryRepository
+     * @param int $page
      * @return Response
      */
-    public function list(BlogRepository $blogRepository, CategoryRepository $categoryRepository): Response
+    public function list(BlogRepository $blogRepository, CategoryRepository $categoryRepository, int $page): Response
     {
         return $this->render('blog/list.html.twig', [
-            'posts' => $blogRepository->findAll(),
+            'posts_paginated' => $blogRepository->findAllPaginated($page),
             'categories' => $categoryRepository->findBy(['environnement' => '2'])
         ]);
     }
@@ -47,7 +49,7 @@ class BlogController extends AbstractController
     }
 
     /**
-     * @Route("/blog/{slug}", name="blog_show_detail", methods={"GET"})
+     * @Route("/blog/show/{slug}", name="blog_show_detail", methods={"GET"})
      * @param Blog $blog
      * @return Response
      */
@@ -63,15 +65,18 @@ class BlogController extends AbstractController
             'blog' => $blog,
         ]);
     }
+
     /**
      * @Route("/admin/blog", name="blog_index", methods={"GET"})
+     * @Route("/admin/blog/{page<[1-9]\d*>}", methods="GET", name="blog_index_paginated")
      * @param BlogRepository $blogRepository
+     * @param int $page
      * @return Response
      */
-    public function index(BlogRepository $blogRepository): Response
+    public function index(BlogRepository $blogRepository, int $page): Response
     {
         return $this->render('blog/index.html.twig', [
-            'blogs' => $blogRepository->findAll(),
+            'blogs_paginate' => $blogRepository->findAllPaginated($page),
         ]);
     }
 
@@ -89,13 +94,13 @@ class BlogController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $documentsCollection = [$form->getData()->getMainImage()[0]];
-            // Fais pas attention a ce foreach
             foreach ($documentsCollection as $key => $result){
                 if ($result) {
                     $data = $fileService->transformToWebP($result->getFile());
                     $result->setCompleteUrl($data['filename']);
                     $result->setFolder($data['folder']);
                     $result->setBlog($blog);
+                    $result->setExt('.webp');
                     $result->setUpdatedAt(new DateTime('now'));
                     $blog->addMainImage($result);
                     $fileService->moveToFolderAndModifyToWebP($this->getParameter($data['folder']), $data['ext'], $data['filename']);
@@ -109,7 +114,7 @@ class BlogController extends AbstractController
             $entityManager->persist($blog);
             $entityManager->flush();
 
-            return $this->redirectToRoute('blog_index');
+            return $this->redirectToRoute('blog_index_paginated', ['page' => 1]);
         }
 
         return $this->render('blog/new.html.twig', [
@@ -139,6 +144,7 @@ class BlogController extends AbstractController
                         $image->setCompleteUrl($dataEdit['filename']);
                         $image->setBlog($blog);
                         $image->setFolder('images');
+                        $image->setExt('.webp');
                         $blog->addMainImage($image);
                         if ($image->getTempFileName()) {
                             $fileService->uploadFolder($this->getParameter($dataEdit['folder']), $dataEdit['ext'], $dataEdit['filename'], $image->getTempFileName().'.webp');
@@ -163,7 +169,7 @@ class BlogController extends AbstractController
     }
 
     /**
-     * @Route("/admin/blog/{id}", name="blog_delete", methods={"DELETE"})
+     * @Route("/admin/blog/delete/{id}", name="blog_delete", methods={"DELETE"})
      * @param Request $request
      * @param Blog $blog
      * @return Response
@@ -171,15 +177,17 @@ class BlogController extends AbstractController
     public function delete(Request $request, Blog $blog): Response
     {
         if ($this->isCsrfTokenValid('delete'.$blog->getId(), $request->request->get('_token'))) {
-            $image = $this->getParameter($blog->getMainImage()[0]->getFolder()).'/'.$blog->getMainImage()[0]->getCompleteUrl();
-            if (file_exists($image.'.webp')) {
-                unlink($image.'.webp');
+            if ($blog->getMainImage()[0] !== null) {
+                $image = $this->getParameter($blog->getMainImage()[0]->getFolder()).'/'.$blog->getMainImage()[0]->getCompleteUrl();
+                if (file_exists($image.'.webp')) {
+                    unlink($image.'.webp');
+                }
             }
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($blog);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('blog_index');
+        return $this->redirectToRoute('blog_index_paginated', ['page' => 1]);
     }
 }
